Added html code to add new property, edit and delete a property, delete image to Internal_properties view file.
I also created model to get properties from DB. Then the model is called
in FindEstate controller and it populates the findEstate view with
property data. Next I need to put it into the html table. And of course
to make work the property search --> only corresponding properties
should display in results.